案件信息 餐馆信息功能大致实现 导入导出和上传附件功能实现

+++ src/main/java/com/example/bishe/config/base/AbstractPaginationQueryCondition.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class AbstractPaginationQueryCondition {
-    private Double start;
-    private Double limit;
-    private Double page;
+    private Integer start;
+    private Integer limit;
+    private Integer page;
 }
